// src/components/component-types.d.ts
// This file was generated by the storyblok CLI.
// DO NOT MODIFY THIS FILE BY HAND.
import type { ISbStoryData } from "storyblok";
export interface CarouselStoryblokType {
  carousel_items: (
    | CarouselStoryblokType
    | FeatureStoryblokType
    | GridStoryblokType
    | TeaserStoryblokType
  )[];
  slideInterval?: string;
  indicators?: boolean;
  pauseOnHover?: boolean;
  component: "carousel";
  _uid: string;
  [k: string]: any;
}

export interface FeatureStoryblokType {
  name?: string;
  component: "feature";
  _uid: string;
  [k: string]: any;
}

export interface GridStoryblokType {
  columns?: (
    | CarouselStoryblokType
    | FeatureStoryblokType
    | GridStoryblokType
    | TeaserStoryblokType
  )[];
  component: "grid";
  _uid: string;
  [k: string]: any;
}

export interface PageStoryblokType {
  header?: (
    | CarouselStoryblokType
    | FeatureStoryblokType
    | GridStoryblokType
    | TeaserStoryblokType
  )[];
  headline: string;
  main?: (
    | CarouselStoryblokType
    | FeatureStoryblokType
    | GridStoryblokType
    | TeaserStoryblokType
  )[];
  footer?: (
    | CarouselStoryblokType
    | FeatureStoryblokType
    | GridStoryblokType
    | TeaserStoryblokType
  )[];
  component: "page";
  _uid: string;
  [k: string]: any;
}

export interface TeaserStoryblokType {
  headline?: string;
  component: "teaser";
  _uid: string;
  [k: string]: any;
}